fix(template): only delete file if it exists
Add a check to the `auto_delete_file_on_delete` function in `models.py`
to only delete the template file if it exists.

+++ document_merge_service/api/models.py
@@ -54,7 +54,8 @@ class Template(models.Model):
 def auto_delete_file_on_delete(sender, instance, **kwargs):
     """Delete template file from filesystem when `Template` object is deleted."""
 
-    DefaultStorage().delete(instance.template.name)
+    if instance.template:
+        DefaultStorage().delete(instance.template.name)
 
 
 @receiver(models.signals.pre_save, sender=Template)
